Bug fix for within profile image upload

+++ users/views.py
@@ -27,7 +27,8 @@ def profile(request):
         if u_form.is_valid() and p_form.is_valid():
             u_form.save()
             if os.path.exists(oldImgPath):
-                os.remove(oldImgPath)
+                if 'default.jpg' not in oldImgPath:
+                    os.remove(oldImgPath)
             p_form.save()
             messages.success(request, f'Your account has been updated!')
             return redirect('profile')
